[ REFACTOR ] 판매점 필터 선택 시, 해당 필터 적용
